Move idtoken to private method in session cookie
